feat: populate DNS cache
Without making a request with the socks proxy enabled globally, the dns for the given domain may not resolve.

Make a request to that domain in order to populate firefox's dns cache, so subsequent requests can use that cache.

It should be noted eventfully this cache may expire. We may need to add some sort of timer and a check to re-populate the cache on some interval. Would rather avoid that due to global proxy needing to be set while we make the request.